requirements.txt
opencv-python>=4.11.0
realesrgan-ncnn-py>=2.0.0
pillow>=11.2.1
numpy>=2.2.5
scikit-image>=0.18.0
